dp: n2193 오답풀이 수정(overflow 따지기)

+++ src/baekjoon/dp/n2193.java
@@ -1,6 +1,7 @@
 package baekjoon.dp;
 import java.io.*;
 
+/*오답 이유: overflow 확인 안 함..경계값 당연하게 대입해봐야 함*/
 public class n2193 {
 
 	public static void main(String[] args) throws IOException{
@@ -10,7 +11,7 @@ public class n2193 {
 		/*
 		 * if(n==1) { System.out.println(1); return; }
 		 */
-		int[][] dp = new int[2][n+1];
+		long[][] dp = new long[2][n+1];
 		dp[0][1] = 0; dp[1][1] = 1;//1만 존재
 		for(int i=2; i<n+1; i++) {
 			dp[0][i] = dp[0][i-1] + dp[1][i-1];
